Move exceptions out of `error` packages
