+++ client/src/main/java/org/spine3/type/KnownTypes.java
@@ -62,7 +62,6 @@ import com.google.protobuf.Value;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.spine3.annotations.Internal;
-import org.spine3.type.error.UnknownTypeException;
 
 import javax.annotation.Nullable;
 import java.lang.reflect.InvocationTargetException;
