+++ client/src/main/java/org/spine3/type/TypeUrl.java
@@ -37,7 +37,6 @@ import org.spine3.base.Event;
 import org.spine3.envelope.CommandEnvelope;
 import org.spine3.envelope.EventEnvelope;
 import org.spine3.envelope.MessageEnvelope;
-import org.spine3.type.error.UnknownTypeException;
 
 import java.util.List;
 import java.util.Objects;
