+++ client/src/test/java/org/spine3/protobuf/error/UnknownTypeExceptionShould.java
@@ -21,7 +21,7 @@
 package org.spine3.protobuf.error;
 
 import org.junit.Test;
-import org.spine3.type.error.UnknownTypeException;
+import org.spine3.type.UnknownTypeException;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
