+++ server/src/main/java/org/spine3/server/aggregate/MissingAggregateIdException.java
@@ -17,21 +17,36 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.spine3.server.aggregate.error;
+package org.spine3.server.aggregate;
 
 import com.google.protobuf.Message;
+import org.spine3.base.Identifiers;
 
 /**
- * This exception is thrown on a discovery of an event class, which is not handled by any of
- * the applier methods of an aggregate class.
+ * Exception is thrown if a command, which is intended to be used for an aggregate,
+ * does not have {@code getAggregateId()} method.
+ *
+ * <p>To have this method in Java, corresponding Protobuf message definition must have
+ * the property called {@code aggregate_id}.
  *
  * @author Mikhail Melnik
+ * @author Alexander Yevsyukov
  */
-public class MissingEventApplierException extends RuntimeException {
+public class MissingAggregateIdException extends RuntimeException {
 
     private static final long serialVersionUID = 0L;
 
-    public MissingEventApplierException(Message event) {
-        super("There is no registered applier for the event: " + event.getClass());
+    public MissingAggregateIdException(Message command, String methodName, Throwable cause) {
+        super(createMessage(command, methodName), cause);
+    }
+
+    private static String createMessage(Message command, String methodName) {
+        return "Unable to call ID accessor method " + methodName + " from the command of class "
+                + command.getClass().getName();
+    }
+
+    public MissingAggregateIdException(String commandClassName, String propertyName) {
+        super("The first property of the aggregate command " + commandClassName +
+                " must define aggregate ID with a name ending with '" + Identifiers.ID_PROPERTY_SUFFIX + "'. Found: " + propertyName);
     }
 }
