+++ server/src/main/java/org/spine3/server/aggregate/MissingEventApplierException.java
@@ -17,12 +17,21 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+package org.spine3.server.aggregate;
+
+import com.google.protobuf.Message;
 
 /**
- * This package contains exceptions related to type identification.
+ * This exception is thrown on a discovery of an event class, which is not handled by any of
+ * the applier methods of an aggregate class.
+ *
+ * @author Mikhail Melnik
  */
+public class MissingEventApplierException extends RuntimeException {
 
-@ParametersAreNonnullByDefault
-package org.spine3.type.error;
+    private static final long serialVersionUID = 0L;
 
-import javax.annotation.ParametersAreNonnullByDefault;
+    public MissingEventApplierException(Message event) {
+        super("There is no registered applier for the event: " + event.getClass());
+    }
+}
