+++ server/src/main/java/org/spine3/server/command/CommandException.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.server.command.error;
+package org.spine3.server.command;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.protobuf.Message;
