+++ server/src/main/java/org/spine3/server/command/CommandStore.java
@@ -26,7 +26,6 @@ import org.spine3.base.CommandStatus;
 import org.spine3.base.Error;
 import org.spine3.base.Errors;
 import org.spine3.base.Failure;
-import org.spine3.server.command.error.CommandException;
 
 import java.util.Iterator;
