+++ server/src/main/java/org/spine3/server/event/UnsupportedEventException.java
@@ -17,7 +17,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.spine3.server.event.error;
+package org.spine3.server.event;
 
 import com.google.protobuf.Message;
 import org.spine3.base.Error;
