+++ server/src/main/java/org/spine3/server/transport/package-info.java
@@ -19,7 +19,8 @@
  */
 
 /**
- * This package contains classes and interfaces defining data transport options between Spine apps and the outer world.
+ * This package contains classes and interfaces defining data transport options between Spine apps
+ * and the outer world.
  *
  * <p>Typical transport options include gRPC servers, Firebase adapters etc.
  */
