+++ server/src/test/java/org/spine3/server/CommandServiceShould.java
@@ -31,7 +31,7 @@ import org.spine3.base.Commands;
 import org.spine3.base.Response;
 import org.spine3.base.Responses;
 import org.spine3.server.command.CommandBus;
-import org.spine3.server.command.error.UnsupportedCommandException;
+import org.spine3.server.command.UnsupportedCommandException;
 import org.spine3.server.transport.GrpcContainer;
 import org.spine3.testdata.TestCommandBusFactory;
