+++ server/src/test/java/org/spine3/server/command/CommandBusShould.java
@@ -30,9 +30,6 @@ import org.spine3.base.Error;
 import org.spine3.base.Response;
 import org.spine3.envelope.CommandEnvelope;
 import org.spine3.envelope.MessageEnvelope;
-import org.spine3.server.command.error.CommandException;
-import org.spine3.server.command.error.InvalidCommandException;
-import org.spine3.server.command.error.UnsupportedCommandException;
 import org.spine3.server.users.CurrentTenant;
 import org.spine3.test.Tests;
 import org.spine3.test.command.AddTask;
