+++ server/src/test/java/org/spine3/server/command/CommandBusShouldHandleCommandStatus.java
@@ -61,7 +61,7 @@ import static org.spine3.base.Commands.getId;
 import static org.spine3.base.Commands.getMessage;
 import static org.spine3.base.Commands.setSchedule;
 import static org.spine3.protobuf.Values.newStringValue;
-import static org.spine3.server.command.error.CommandExpiredException.commandExpiredError;
+import static org.spine3.server.command.CommandExpiredException.commandExpiredError;
 import static org.spine3.test.TimeTests.Past.minutesAgo;
 
 @SuppressWarnings({"ClassWithTooManyMethods", "OverlyCoupledClass"})
